Rename package wordmaps to mappings
