+++ internal/mappings/csv.go
@@ -1,4 +1,4 @@
-package wordmaps
+package mappings
 
 import (
 	"github.com/ericcornelissen/stringsx"
