+++ internal/mappings/markdown.go
@@ -1,4 +1,4 @@
-package wordmaps
+package mappings
 
 import (
 	"regexp"
