+++ internal/mappings/markdown_fuzz.go
@@ -1,6 +1,6 @@
 // +build gofuzz
 
-package wordmaps
+package mappings
 
 func FuzzMarkDown(data []byte) int {
 	s := string(data)
